# swagger-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Ghar Ka Sathi Production APIs",
    "version": "1.0.0",
    "description": "Comprehensive interactive API documentation for customer, expert, and admin application workflows. Built to support 80+ endpoints."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080/api/v1",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/address": {
      "post": {
        "tags": ["4. Address Management"],
        "summary": "Add Address to Book",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "house": { "type": "string", "example": "Flat 302" },
                  "building": { "type": "string", "example": "Green Gardens" },
                  "landmark": { "type": "string", "example": "Opp petrol pump" },
                  "latitude": { "type": "number", "example": 28.5355 },
                  "longitude": { "type": "number", "example": 77.3910 },
                  "tag": { "type": "string", "example": "Home" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Saved" }
        }
      },
      "get": {
        "tags": ["4. Address Management"],
        "summary": "Get All Address Book",
        "responses": {
          "200": { "description": "Address items array" }
        }
      }
    },
...
    "/bookings": {
      "post": {
        "tags": ["9. Bookings Lifecycle"],
        "summary": "Create new service booking",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "categoryId": { "type": "string", "example": "cat_plumber" },
                  "subServiceId": { "type": "string", "example": "sub_pl_leak" },
                  "dateTime": { "type": "string", "example": "2026-07-25T10:00:00Z" },
                  "timeSlot": { "type": "string", "example": "10:00 AM - 12:00 PM" },
                  "address": { "type": "string", "example": "Sector 15 Noida" },
                  "problemDescription": { "type": "string", "example": "Kitchen pipeline leaks" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Created" }
        }
      },
      "get": {
        "tags": ["9. Bookings Lifecycle"],
        "summary": "Get Bookings History (Filtered)",
        "parameters": [
          { "name": "filter", "in": "query", "schema": { "type": "string", "example": "Pending" } }
        ],
        "responses": {
          "200": { "description": "Bookings array" }
        }
      }
    },
...
    "/support/ticket": {
      "post": {
        "tags": ["19. Help & Support Desk"],
        "summary": "Raise customer support ticket",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "subject": { "type": "string", "example": "Wrong invoice" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Raised" }
        }
      }
    },
...
    "/support/tickets/{id}": {
      "get": {
        "tags": ["19. Help & Support Desk"],
        "summary": "Get Support Ticket Details & Replies",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Details" }
        }
      }
    },
    "/support/tickets/{id}/reply": {
      "post": {
        "tags": ["19. Help & Support Desk"],
        "summary": "Reply/Add comment to support ticket",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": { "type": "string", "example": "I did not receive reply yet" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Reply added" }
        }
      }
    },
    "/refund": {
      "post": {
        "tags": ["19. Help & Support Desk"],
        "summary": "Raise Refund Dispute request",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "bookingId": { "type": "string", "example": "bk_123" },
                  "reason": { "type": "string", "example": "Expert did not show up" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Dispute registered" }
        }
      }
    },
    "/report/expert": {
      "post": {
        "tags": ["19. Help & Support Desk"],
        "summary": "Report Expert partner misconduct",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "providerId": { "type": "string", "example": "prov_rajesh" },
                  "reason": { "type": "string", "example": "Overcharged cash payment" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Misconduct logged" }
        }
      }
    },
...
  }
}